Calculate genome 'novelty' from a list of selection sets returned by pick_derep_sets

Calculate genome 'novelty' from a list of selection sets returned by pick_derep_sets

Usage

calculate_novelty(selection_set_results)

Arguments

selection_set_results

a tibble of selection sets

Value

a dataframe with the novelty scores of all genomes in the list 'Novelty' score for each genome is (number of selections) / median rank of selection

Examples

sets <- pick_derep_sets(example_pangenome_matrix)
genome_novelty <- calculate_novelty(sets)
genome_novelty
#> # A tibble: 26 × 8
#>    asm_acc   median_rank number_selections best_…¹ worst…² novel…³ log_n…⁴  RANK
#>    <chr>           <dbl>             <int>   <int>   <int>   <dbl>   <dbl> <int>
#>  1 genome_55         2                  15       2       4   7.5     2.01      1
#>  2 genome_19         3                  10       3       4   3.33    1.20      2
#>  3 genome_31         2.5                 8       2       4   3.2     1.16      3
#>  4 genome_72         3.5                 6       2       5   1.71    0.539     4
#>  5 genome_11         3                   5       2       5   1.67    0.511     5
#>  6 genome_8          2                   3       2       4   1.5     0.405     6
#>  7 genome_85         4                   5       2       5   1.25    0.223     7
#>  8 genome_20         3                   3       2       5   1       0         8
#>  9 genome_28         3                   3       3       3   1       0         9
#> 10 genome_60         4.5                 4       3       5   0.889  -0.118    10
#> # … with 16 more rows, and abbreviated variable names ¹​best_rank, ²​worst_rank,
#> #   ³​novelty_score, ⁴​log_novelty
#> # ℹ Use `print(n = ...)` to see more rows