return a Year column containing the earliest year from the available 'date' fields.

return a Year column containing the earliest year from the available 'date' fields.

Usage

extract_earliest_year(PDD_metadata_table)

Arguments

PDD_metadata_table

an ncbi pathogen detection metadata table

Value

returns the a tibble with a Year column,

Examples

extract_earliest_year(klebsiella_example_dat)
#> Warning: NAs introduced by coercion
#> # A tibble: 200 × 2
#>    target_acc      Year
#>    <chr>          <dbl>
#>  1 PDT000030662.1  2013
#>  2 PDT000097209.1  2012
#>  3 PDT000102016.2  2015
#>  4 PDT000131558.1  2006
#>  5 PDT000131577.1  2003
#>  6 PDT000136458.2  2011
#>  7 PDT000140949.1  2015
#>  8 PDT000162952.1  2011
#>  9 PDT000172132.4  2013
#> 10 PDT000207429.1  2010
#> # … with 190 more rows