generate ftp site download urls for all SNP trees containing the provided isolates

generate ftp site download urls for all SNP trees containing the provided isolates

make_SNPtree_urls(organism, data, PDG)

Arguments

organism

a string ie 'Salmonella' or 'Campylobacter' etc

data

a metadata table, must contain the column 'PDS_acc' from merging in the SNP cluster data

PDG

The PDG version the metadata is from.

Value

returns a vector of ftp download urls for each tar.gz file containing the SNP tree info

Examples

make_SNPtree_urls(organism = 'Klebsiella',
 data = klebsiella_example_dat, PDG = 'PDG000000012.1053')
#> 78 Isolates in the collection are not represented in SNP trees
#>   [1] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074965.1.tar.gz"  
#>   [2] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045327.14.tar.gz" 
#>   [3] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000046982.1.tar.gz"  
#>   [4] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000098799.1.tar.gz"  
#>   [5] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000070365.1.tar.gz"  
#>   [6] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000076332.4.tar.gz"  
#>   [7] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000015204.33.tar.gz" 
#>   [8] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000075205.1.tar.gz"  
#>   [9] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000005547.4.tar.gz"  
#>  [10] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000098760.1.tar.gz"  
#>  [11] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000060404.1.tar.gz"  
#>  [12] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000012992.6.tar.gz"  
#>  [13] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000015678.105.tar.gz"
#>  [14] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000048063.60.tar.gz" 
#>  [15] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000041844.1.tar.gz"  
#>  [16] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000012133.1.tar.gz"  
#>  [17] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000053028.3.tar.gz"  
#>  [18] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000098780.1.tar.gz"  
#>  [19] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074945.1.tar.gz"  
#>  [20] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045338.14.tar.gz" 
#>  [21] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045386.1.tar.gz"  
#>  [22] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000053031.3.tar.gz"  
#>  [23] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074963.1.tar.gz"  
#>  [24] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000090684.2.tar.gz"  
#>  [25] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000092788.1.tar.gz"  
#>  [26] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000077015.9.tar.gz"  
#>  [27] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045421.13.tar.gz" 
#>  [28] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000076318.2.tar.gz"  
#>  [29] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074980.1.tar.gz"  
#>  [30] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000062384.1.tar.gz"  
#>  [31] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000065465.1.tar.gz"  
#>  [32] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074864.1.tar.gz"  
#>  [33] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000046973.2.tar.gz"  
#>  [34] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000036263.5.tar.gz"  
#>  [35] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000040691.3.tar.gz"  
#>  [36] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000096677.1.tar.gz"  
#>  [37] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000084738.1.tar.gz"  
#>  [38] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000042796.1.tar.gz"  
#>  [39] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000005629.4.tar.gz"  
#>  [40] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000040457.7.tar.gz"  
#>  [41] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000065281.5.tar.gz"  
#>  [42] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000056111.12.tar.gz" 
#>  [43] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000095229.5.tar.gz"  
#>  [44] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000093261.1.tar.gz"  
#>  [45] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000075287.1.tar.gz"  
#>  [46] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045531.1.tar.gz"  
#>  [47] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000070357.2.tar.gz"  
#>  [48] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000042766.1.tar.gz"  
#>  [49] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000009754.59.tar.gz" 
#>  [50] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074884.3.tar.gz"  
#>  [51] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045521.2.tar.gz"  
#>  [52] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045542.3.tar.gz"  
#>  [53] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000052366.22.tar.gz" 
#>  [54] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000060597.1.tar.gz"  
#>  [55] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000079409.19.tar.gz" 
#>  [56] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074459.2.tar.gz"  
#>  [57] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074906.2.tar.gz"  
#>  [58] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000085647.1.tar.gz"  
#>  [59] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000042789.1.tar.gz"  
#>  [60] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045411.1.tar.gz"  
#>  [61] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000033027.1.tar.gz"  
#>  [62] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000075041.4.tar.gz"  
#>  [63] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000005616.2.tar.gz"  
#>  [64] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000055775.1.tar.gz"  
#>  [65] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000036221.1.tar.gz"  
#>  [66] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000077167.1.tar.gz"  
#>  [67] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000076562.4.tar.gz"  
#>  [68] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045377.11.tar.gz" 
#>  [69] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000055024.5.tar.gz"  
#>  [70] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000072923.1.tar.gz"  
#>  [71] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000075206.7.tar.gz"  
#>  [72] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045274.18.tar.gz" 
#>  [73] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000090692.1.tar.gz"  
#>  [74] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000036943.4.tar.gz"  
#>  [75] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000050952.1.tar.gz"  
#>  [76] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000041726.10.tar.gz" 
#>  [77] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000042778.1.tar.gz"  
#>  [78] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000005975.9.tar.gz"  
#>  [79] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000098756.1.tar.gz"  
#>  [80] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000097794.1.tar.gz"  
#>  [81] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045465.5.tar.gz"  
#>  [82] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000085573.3.tar.gz"  
#>  [83] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045260.7.tar.gz"  
#>  [84] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045318.3.tar.gz"  
#>  [85] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000075309.1.tar.gz"  
#>  [86] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000005543.11.tar.gz" 
#>  [87] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000090149.1.tar.gz"  
#>  [88] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000075211.1.tar.gz"  
#>  [89] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045562.1.tar.gz"  
#>  [90] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000070360.1.tar.gz"  
#>  [91] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074857.2.tar.gz"  
#>  [92] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000047783.23.tar.gz" 
#>  [93] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000090691.1.tar.gz"  
#>  [94] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000085658.2.tar.gz"  
#>  [95] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074956.1.tar.gz"  
#>  [96] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045679.3.tar.gz"  
#>  [97] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000098808.1.tar.gz"  
#>  [98] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000017083.14.tar.gz" 
#>  [99] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045323.8.tar.gz"  
#> [100] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000065420.2.tar.gz"  
#> [101] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000017392.9.tar.gz"  
#> [102] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074793.3.tar.gz"  
#> [103] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074832.1.tar.gz"  
#> [104] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000012123.2.tar.gz"  
#> [105] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000076556.2.tar.gz"  
#> [106] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000077827.1.tar.gz"