make specific ftp download paths for a dataframe with ftp_paths and assembly accessions

make specific ftp download paths for a dataframe with ftp_paths and assembly accessions

make_download_urls(data, type)

Arguments

data

a dataframe with the columns 'ftp_path' and 'asm_acc' ftp_path should be a column produced by the function make_ftp_paths()

type

type of download path to generate, one of: 'fna', 'gbff', 'gff', 'gtf', 'faa', 'cds'

Value

returns the original dataframe with an added column, named "type_download"

Examples

make_download_urls(klebsiella_example_dat, type='fna')
#> # A tibble: 200 × 70
#>    `#label`  FDA_lab_id HHS_region IFSAC_category LibraryLayout PFGE_PrimaryEnz…
#>    <chr>     <chr>      <chr>      <chr>          <chr>         <chr>           
#>  1 PDT00019… NULL       NULL       NULL           NULL          NULL            
#>  2 PDT00076… NULL       NULL       NULL           NULL          NULL            
#>  3 PDT00041… NULL       NULL       NULL           NULL          NULL            
#>  4 PDT00100… NULL       NULL       NULL           NULL          NULL            
#>  5 PDT00122… NULL       NULL       NULL           NULL          NULL            
#>  6 PDT00094… NULL       NULL       NULL           NULL          NULL            
#>  7 PDT00121… NULL       NULL       NULL           NULL          NULL            
#>  8 PDT00100… NULL       NULL       NULL           NULL          NULL            
#>  9 PDT00084… NULL       NULL       NULL           NULL          NULL            
#> 10 PDT00037… NULL       NULL       NULL           NULL          NULL            
#> # … with 190 more rows, and 64 more variables:
#> #   PFGE_SecondaryEnzyme_pattern <chr>, Platform <chr>, Run <chr>,
#> #   asm_acc <chr>, asm_level <chr>, asm_stats_contig_n50 <dbl>,
#> #   asm_stats_length_bp <dbl>, asm_stats_n_contig <dbl>, assembly_method <chr>,
#> #   attribute_package <chr>, bioproject_acc <chr>, bioproject_center <chr>,
#> #   biosample_acc <chr>, isolate_identifiers <chr>, collected_by <chr>,
#> #   collection_date <chr>, epi_type <chr>, fullasm_id <chr>, …