generate ftp site download urls for all SNP trees containing the provided isolates

generate ftp site download urls for all SNP trees containing the provided isolates

make_SNPtree_urls(organism, data, PDG)

Arguments

organism

a string ie 'Salmonella' or 'Campylobacter' etc

data

a metadata table, must contain the column 'PDS_acc' from merging in the SNP cluster data

PDG

The PDG version the metadata is from.

Value

returns a vector of ftp download urls for each tar.gz file containing the SNP tree info

Examples

make_SNPtree_urls(organism = 'Klebsiella',
 data = klebsiella_example_dat, PDG = 'PDG000000012.1053')
#> 86 Isolates in the collection are not represented in SNP trees
#>  [1] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000012469.1.tar.gz"  
#>  [2] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000048063.68.tar.gz" 
#>  [3] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000098771.3.tar.gz"  
#>  [4] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000103219.1.tar.gz"  
#>  [5] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000052089.14.tar.gz" 
#>  [6] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000072037.3.tar.gz"  
#>  [7] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000095581.1.tar.gz"  
#>  [8] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000074971.1.tar.gz"  
#>  [9] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000012621.3.tar.gz"  
#> [10] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000095229.11.tar.gz" 
#> [11] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000103191.1.tar.gz"  
#> [12] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000052366.23.tar.gz" 
#> [13] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000038612.1.tar.gz"  
#> [14] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000103213.1.tar.gz"  
#> [15] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000096677.2.tar.gz"  
#> [16] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000036296.3.tar.gz"  
#> [17] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000036263.5.tar.gz"  
#> [18] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000085573.4.tar.gz"  
#> [19] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000095239.2.tar.gz"  
#> [20] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000052185.1.tar.gz"  
#> [21] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045319.1.tar.gz"  
#> [22] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000012628.23.tar.gz" 
#> [23] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000012112.56.tar.gz" 
#> [24] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000077015.12.tar.gz" 
#> [25] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000015418.30.tar.gz" 
#> [26] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045299.5.tar.gz"  
#> [27] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000012124.1.tar.gz"  
#> [28] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000015678.108.tar.gz"
#> [29] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000070361.1.tar.gz"  
#> [30] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000102667.2.tar.gz"  
#> [31] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000060185.1.tar.gz"  
#> [32] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000040466.3.tar.gz"  
#> [33] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045368.1.tar.gz"  
#> [34] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045297.24.tar.gz" 
#> [35] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045394.20.tar.gz" 
#> [36] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000041859.6.tar.gz"  
#> [37] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045360.10.tar.gz" 
#> [38] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000084730.1.tar.gz"  
#> [39] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000012626.6.tar.gz"  
#> [40] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000058604.5.tar.gz"  
#> [41] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000096689.1.tar.gz"  
#> [42] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000006578.57.tar.gz" 
#> [43] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000055633.9.tar.gz"  
#> [44] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000040472.2.tar.gz"  
#> [45] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045443.1.tar.gz"  
#> [46] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000054005.3.tar.gz"  
#> [47] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000036176.1.tar.gz"  
#> [48] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000019488.9.tar.gz"  
#> [49] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000046936.3.tar.gz"  
#> [50] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000094136.1.tar.gz"  
#> [51] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000079862.7.tar.gz"  
#> [52] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000005616.2.tar.gz"  
#> [53] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045272.11.tar.gz" 
#> [54] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000101843.1.tar.gz"  
#> [55] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000085669.1.tar.gz"  
#> [56] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000048069.3.tar.gz"  
#> [57] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000038608.2.tar.gz"  
#> [58] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000065414.2.tar.gz"  
#> [59] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045352.23.tar.gz" 
#> [60] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000038923.3.tar.gz"  
#> [61] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000078629.4.tar.gz"  
#> [62] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000075041.5.tar.gz"  
#> [63] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000066129.3.tar.gz"  
#> [64] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000101833.1.tar.gz"  
#> [65] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045410.6.tar.gz"  
#> [66] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000050839.3.tar.gz"  
#> [67] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000050841.5.tar.gz"  
#> [68] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000036246.5.tar.gz"  
#> [69] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000036336.23.tar.gz" 
#> [70] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000103169.1.tar.gz"  
#> [71] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000053034.5.tar.gz"  
#> [72] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000103224.1.tar.gz"  
#> [73] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045421.15.tar.gz" 
#> [74] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000018683.4.tar.gz"  
#> [75] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000084736.1.tar.gz"  
#> [76] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000075219.3.tar.gz"  
#> [77] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045135.6.tar.gz"  
#> [78] "https://ftp.ncbi.nlm.nih.gov/pathogen/Results/Klebsiella/PDG000000012.1053/SNP_trees/PDS000045371.1.tar.gz"